Added Gmaps as background to app
Using Gmap, default SF civic center as locaiton. In initializing, getting the
current location with browser and setting the map to center there. If no
location found, no centering of map is done.

CSS is done by having overlay class to overlay logo on top of map with position
absolute. The footer is not set to absolute and floats on bottom.

TODO:
- Include a marker of your current location that you can move around.
- Send the location data based on marker rather then browser navigator
position like now.